[PCF-710]🔨 Core and mocks compose configs use the dev Dockerfile and start direclty  when docker compose up
We extended the Dockerfile to all mocks and the core.
Now, the small app no longer uses _start-all
and the packages are installed via the Dockerfile instead of the _up command.

For now, we only use hardcoded commands in production mode.
This is temporary; in one of the next steps, we will need to implement
a way to flag the command as either development or production mode.

We also keep in mind, that we have removed the pm2 logs command,
We will have to evaluate later the situation of the logs.