[PCF-710]🔥 Remove start from dks _switch and _up does not install anymore npm packages
This cleanup ensures that the POC is working.
We no longer use _start-all.
Additionally, we can be sure that packages are now
installed via the Dockerfile instead of the _up command.

In the next steps, we will need to adapt the code
of the other stacks (medium, hybridge, etc.) to ensure that
no other app requires the _start-all step anymore.