[PCF-710]🔨 Use node_modules volumes in small stack
In this commit, we removed the current user
from the node-modules-installer
to allow the installation of node_modules
in the node_modules volume.

Adding this volume changes the ownership
of the node_modules folder to root.
So it is not possible anymore to write in this
folder as the current user.
We removed the current user,
and the Docker default user root
is now used instead,
which allows us to write to this folder.

Then we updated the compose files for
fsa1, fia1, and core to test the small stack.
These three apps share the same node_modules volume,
allowing a single installation
for all the Nest apps in the backend.

The other files are copied using a mounted volume,
with the owner set to the current user.
However, the copied node_modules folder
is overwritten by the volume with root as owner.

While fia1 and fsa2 seem to
work correctly with this commit,
the core app is not functioning properly.