refactor: code review
Promise.all is a red flag when applied to unconstrained collections (here, the size of rawDataProviders is unbounded) because it implies that an unbounded number of promises have been created, which in turn means that a possibly large number of asynchronous tasks have been launched simultaneously.

It's generally better for performance and memory to process items sequentially (as above thanks to for + await), or with controlled concurrency by using something like https://github.com/sindresorhus/p-map which allows limiting the number of promises created concurrently.

In this particular case it seems the validate calls will return synchronously in practice, so it's very likely that sequential processing is the best option in terms of overhead.