🔨  Move ci admin tests in federation
In this PR, we aim to modify only the necessary files to move admin into federation. We continue to use symlinks. The only required adjustments for the tests are some paths and how the symlinks are built. We no longer use the federation-admin repository, so the clone step has been removed.