🐛 Add prestart when launching admin and correct no css in dev
By moving admin into federation, I realized there's a bug introduced
in a previous PR when starting admin in development mode.

The prestart command is not called, which prevents the app from building correctly
when using only the start:dev script — resulting in missing CSS, JS, images, etc.

I took the opportunity in this PR to fix it by including the prestart command in start:dev.