🐛 introspection sub is inaccurate
The oidc-provider library expects the accountId to match the user’s sub
when generating introspection responses. Our current implementation
sets accountId to the sessionId, which differs from the sub.

I propose introducing a session alias that maps the session to the
accountId. This leverages existing mechanisms and minimizes
implementation overhead.