🔨 Start dev stack with the Back Dockerfile used in prod
This is the first milestone toward simplifying the stack
and enabling apps to be launched in DKS with a single command: docker compose up.
To use the same Dockerfile in development as in production,
I created an initial version of a new stage in the Dockerfile that supports Bash.

This approach allows me to minimize changes to the development stack.
For now, this stage is still quite simple,
but it will be improved in future iterations.

I also added a mechanism to detect which containers are “node” containers
that need to be restarted, by almost hardcoding the names of the existing apps.

I’m keeping the previous parent-ancestor mechanism for now,
just to be safe in case I forgot a special case.

Some apps still use the old Dockerfile (logs and hybridge apps).
The migration to the new Dockerfile will be done
in a dedicated PR for these apps,
to keep this PR as slim as possible and focused on the core and the mock.