Make mongo-fca-low automatically load its default contents
This removes another dependency on docker-stack. I've left the
call to `_reset_mongodb` in `up.sh` though because removing it
would change the behavior — currently every `up` causes the
database to be reset, whereas the `initdb.d` contents are only
executed when initially creating the database (which will only
currently happen after a `compose down --volumes` or during
a `compose up --renew-anon-volumes`). If this behavior change
is fine though, the call in `up.sh` can be removed later.