Use node_modules from node image directly instead of copying to an anonymous volume
The previous setup of:

    volumes:
      - '../../../back:/var/www/app'
      - '/var/www/app/node_modules'

Meant that `/var/www/app/node_modules` was first shadowed by the bind mount
of `back`, then "re-introduced" as an anonymous volume whose contents were
automatically copied from the image (see
https://docs.docker.com/engine/storage/volumes/#mounting-a-volume-over-existing-data).

This unnecessarily duplicated the entire `node_modules` directory into as many anonymous
volumes as there were containers, and also took several seconds at container creation.Also,
since Docker Compose will preserve anonymous volumes by default, it was easy to end up in
a situation where the underlying image was rebuilt with new `node_modules` content that
was not picked up (as the anonymous volumes contained previous copies).
This also caused the creation of an empty `back/node_modules` directory on the host, as
well as a `back/dist` subdirectory containing the compilation output.

The alternative setup introduced by this commit:
 - copies the entire contents of `back/` into the image;
 - selectively maps a few subdirectories of `back/` into `/var/www/app`, leaving
   `node_modules` directly accessible from its image version. It is expected that
   almost all changes made in development will happen in one of those directories.
 - creates a `tmpfs` mount for the `dist/` subdirectory which is the only place under
   `/var/www/app` where the app is expected to write.

Similar logic was applied to `admin`.

As a result, no changes are made to the host filesystem outside of `docker/volumes`, making
the `_clean` function of `docker-stack` unnecessary (as it was implemented, at least).