🔨 Streamline dev Docker images
Yarn will store its cache by default in `/usr/local/share/.cache`.

Instead of keeping that in the final image, we move it to a cache mount
which gives it an opportunity to be reused by the next build.

We also make /tmp a tmpfs mount because yarn dumps stuff there as well.