🔒️(totp) use RFC-compliant OTP window and simplify testing
Relevant RFC: https://datatracker.ietf.org/doc/html/rfc6238#section-5.2

The TOTP RFC recommends using 'an acceptable OTP transmission delay window'
of at most one. Zero is too tight for practical purposes and leads to wasted
time waiting for the next TOTP token to appear when using an authenticator
app.

In addition, we have two mitigations for this problem in Cypress: originally
the environment variable TOTP_WINDOW was set to "loose" to allow this one-step
delay. However, this could not possibly have worked, since this variable is
only ever seen by the Cypress testing environment, which *generates* OTP codes,
and could not have any effect on the admin application which *accepts* OTP
codes - and the 'window' config only affects acceptance.

Because this was ineffective a second mitigation was added to the tests with
a 5-second wait when the OTP code had less than 5 seconds' validity (increased
from an initial value of 2). This seems to have worked but increases CI
duration, which is not ideal.

Therefore, without compromising security, in this PR we set the delay window
to one time step (30 seconds) in the admin application, and remove both
mitigations from the tests.