🐛(authent) fixes LocalGuard to allow bump to passport 0.7.0
This commit mainly fixes a bug that caused all login attempts to fail
on admin, leading to a nearly 100% failure rate on E2E tests.

The fix consists of adding a missing "await" before super.logIn. In
its absence, the redirect to admin's home page would occur before
the authenticated user was added to the session by passport, and
passport regenerated the session. The order of events matter as
redirect() flushes headers; setting cookies, such as the session
cookie, after the call to redirect(), has no effect.

On version 0.5.3 of Passport, things worked because passport did
not regenerate the session, and LocalAuthGuard regenerated it prior
to the logIn() call.

However starting with 0.6.0, Passport started regenerating the
session:
  https://medium.com/passportjs/fixing-session-fixation-b2b68619c51d

Therefore LocalAuthGuard no longer needs to do that.

The official documentation on NestJs and Passport is incomplete
and somewhat confusing but it seems that we do need to extend
AuthGuard *and* call the logIn method.

https://dev.to/nestjs/setting-up-sessions-with-nestjs-passport-and-redis-210

The unit tests are no longer really necessary, since all they checked
was the call to regenerateSession, which no longer happens, so rather
than keep them for the sake of illusory coverage it seems better to
delete them.