🥅 better handle misconfigured idp
When idp is missconfigured, it can be referenced in fqdnToProviders but
prevented from loading because of a validation error. This result in a
500 unexpected error while trying to access the name of the not loaded
IDP.

I propose to throw an explicit error when this case happen. I propose
some renaming to make it more explicit whether we are using idp or
ipd_id.