🧑‍💻 remove flaky tests (#466)
@Morendil:

I think the intention was to write this:
```
   const itemTest1 = serviceProviderFactory.createServiceProviderFromDb({
     ...defaultSpMock,
     name: 'joker',
   });
```

…but unfortunately we have this
```
   const itemTest1 = serviceProviderFactory.createServiceProviderFromDb({
     name: 'joker',
     ...defaultSpMock,
   });
```

…which doesn't do the same thing at all. Instead of changing the name, we keep the original name `monfs` for both objects. (I checked by outputting the items via `console.debug`.)

As written the unit test works because it compare two strictly identical SPs being sorted in ascending or descending order. Of course since the values are identical, the order doesn't matter… except in some rare cases where the values are created on distinct sides of a millisecond timing boundary.

However, fixing the above problem results in the test failing consistently rather than intermittently - the "fix" makes the test impossible to pass rather than vacuous.

My best idea is to remove the test altogether: no test at all is better than a vacuous test which fails sporadically.

@rdubigny:

Theses test pretend to cover the sorting feature of ServiceProviderController.list but the sorting is made by serviceProviderService.paginate which is mocked.

We can safely remove these tests as they are not testing what they
pretend to test.