♻️  remove session duplication in authorize
We are not sure why the session was duplicated.
It seems intended as a session key rotation to mitigate cookie theft
attacks, but the previous key does not appear to be revoked.

Such protection is relevant when elevating privileges within a session
or when revoking a session using the same key, but that's not the case
in the authorize endpoint. It still applies in the callback endpoint.

This duplication would cause issues when refactoring the prompt to
match the Panva API.

I propose removing it now to make the code clearer.