🔇 remove FQDN in some logs
At the FC_SHOWED_IDP_CHOICE event, fqdn is undefined when no ipd_hint
is provided, which happens in more than 90% of connections. This makes
the field difficult to use.

Additionally, usage of login_hint should be tracked explicitly with a
dedicated event.

I propose removing this as a first step toward the feature that will
skip the interaction page display (corresponding to FC_SHOWED_IDP_CHOICE)
when login_hint is set.