♻️  Full rewrite of tracking lib
Focus on precisely defining the TrackedEventContextInterface type.

Remove FC_REDIRECT_TO_IDP, which is unreliable (not triggered in some
login scenarios), and replace it with the more explicit IDP_CHOSEN.

Retrieve tracking information from the session.

Systematically add the email, as it is a key piece of information and
we have approval to log it.