🗃️  standardize creation of indexes
Problem

The compound index on accountFca.idpIdentityKeys idpUid & idpSub was not
provisionned for the integ and the production environnement.

There is 2 ways to add an index in the project:
- coding it in the mongoose schema
- running a query manually

We should have only one way to create an index.

Proposal

Declare the compound index with mongoose like other existing indexes