🎉 Replace debounce with DNS check
**Problem**
We currently use Debounce to prevent email delivery errors. While this
is useful in PCI, PCF does not send emails.
However, PCF still needs to validate email domains to catch typos like
`gendramerie.interieur.gouv.fr`, which should be blocked instead of
passed to PCI.

Additionally, there are concerns about sending email addresses to
Debounce, as this could lead to data leakage.

**Proposal**
Perform a DNS check on the email domain to ensure it exists.
This approach should be sufficient for our use case.