🥅 Prevent phone number validation from throwing
**Problem**
If an IDP returns the `phone_number` field as an array, the user is
shown a `Y500006 InvalidIdentity` exception.
Since this field is optional, we should ignore the mistyped value
rather than blocking the user.

**Proposal**
If an IDP returns a malformed `phone_number`, set its value to
`undefined` in all cases.