♻️  simplify OIDC Provider error handling 2/3
**Problem**
`OidcProviderErrorService` calls a catchErrorEvents function that adds
a log listener on `oidc-provider` error events.
These listeners calls a `renderError` function that forwards catched
`oidc-provider` errors (wrapped in an `OidcProviderNoWrapperException`)
through a NestJS command bus.
The command bus then triggers an exception filter based on the class
(which is always `OidcProviderBaseRenderedException`).
If the request’s content type is not JSON, (which is the case we are
interested in) the `OidcProviderRenderedHtmlExceptionFilter`’s `catch`
function is called.
This function tries to determine whether the error should trigger a
http redirect by reverse-engineering the error messages from
`oidc-provider`.
If a redirection is needed (which is the case we are intersted in), it
then logs the error and shawdowly let `oidc-provider` handle the
redirection.

**Solution**
Add listeners that log the error and let `oidc-provider` handle the
redirection.