♻️  simplify OIDC Provider error handling 3/3
**Problem**
`OidcProviderErrorService` implements the `renderError` function used
by `oidc-provider` to render errors.
The current implementation forwards `oidc-provider` errors (wrapped in
an `OidcProviderNoWrapperException`) through a NestJS command bus.
The command bus then triggers an exception filter based on the class
(which is always `OidcProviderBaseRenderedException`).

If the request’s content type is not JSON, (which is always the case
when `oidc-provider` calls `renderError`) the
`OidcProviderRenderedHtmlExceptionFilter`’s `catch` function is called.
This function tries to determine whether the error should trigger a
http redirect by reverse-engineering the error messages from
`oidc-provider`.
If no redirection is needed (which is always the case when
`oidc-provider` calls `renderError`), it then calls the
`FcWebHtmlExceptionFilter`’s `catch` method, which logs the exception
again and renders the page.

**Proposal**
The `renderError` function should render a fine error page.