🔥 Remove useless logger plugin
**Problem**
Most of the time, the stacktrace key shows a *Filter.logException* call
which is not useful for debugging and adds a lot of noise.
Plus this plugin is also active in production environment which should
not be the case due to Logger_THRESHOLD set to DEBUG in production environment.

**Proposal**
Remove this plugin solves the first problem.
The second one is to set Logger_THRESHOLD to INFO in production environment.
see: https://github.com/proconnect-gouv/infra-legacy/pull/77