🔊 turn tracked event into warnings
**Problem**
Business events logs info message on the status of an interaction.
Errors are usually logged in the main logger. FC_FQDN_MISMATCH are an
exception to that usage.

Also, like email_not_safe_to_send, this kind of log refer to event that
should be destined to system operators, for an unwanted state that can
continue to be. According to Taco Jan Osinga cheat sheet on log level,
these should be a warning.

see stackoverflow.com/questions/2031163/when-to-use-the-different-log-levels#answer-64806781

**Proposal**

Turn these logs into warnings.