🗑 remove exception occured handler
**Problem**
There is too much code and patterns involved for just a simple need:

- log the error
- display a fine error when possible

**Proposal**
The console is already overrided to log proper json error.
There already is a log listner on oidc-provider error.
There already is a fine error display for error going through
oidc-provider "renderError" method.
We just add a middleware that display a nice error if an error happen in
the koa middleware.

With all this we should be covered and we should no longer need most of
the exception lib code.