⚖ better compliance of amr claim
**Problem**
When an IDP returns an empty amr we default the value to an empty array.

This contravenes with the oidc standard that states:

If a Claim is not returned, that Claim Name SHOULD be omitted from the JSON object representing the Claims; it SHOULD NOT be present with a null or empty string value.

**Proposal**
If a claim is not return, do not default it to empty array.