♻️  merge legacy logger into main logger
**Problem**
We should be able to log everything in stdout to be hostable on kube.
There is no reason to keep maintaining 2 loggers as they have the same
configuration but the output stream.

**Proposal**
Merge the 2 loggers into one. If a file path is provided, logs the
tracked event logs into the file.