🔧 Attempt to fix 190001
**Problem**
I was able to reproduce an error 190001 by triggering a
/interaction/:uid request and cancel it before it ends.
The next call to /interaction/:uid calls with a session id that has been
destroyed by the session.reset() function resulting in a sessionNotFound
error.

**Proposal**
I am trying to resolve the issue by implementing a reset that does not
generate a new session Id on each call.
We must investigate deeper on why we need to reset the session at the
first place though.