📦 Use forked `class-validator` package
**Problem**
They are inheritance issues in `class-validator` (see issue:
https://github.com/typestack/class-validator/issues/633).

We updated a PR to fix this problem:
https://github.com/typestack/class-validator/pull/2641.

However, the chances of it ever being merged are low, as the repository
no longer appears to be actively maintained.

**Proposal**
Use the patched package and import it via an "ugly" github link.
This should push us toward eventually migrating to another validation
library such as Zod.