🔊 Display constructor name alongside error code
**Problem**
`YXXXXXX` error codes (where `X` are digits) are not understandable
without the lookup table that maps each code to a specific error.
This makes it difficult to quickly understand what kind of error we are
dealing with.

**Proposal**
Dispolay the constructor name alongside the error code to five users an
immediate hint about what's happening.