🧑‍💻 `id_token_hint` is optional at logout
**Problem**
`id_token_hint` can be tricky to store for some of our Relying Parties.
If they are unable to include it when triggering a logout, we currently
return a Bad Request error because we treat the parameter as mandatory.
However, the OpenID specification stats that this parameter is
recommended but not required.
Reference: https://openid.net/specs/openid-connect-rpinitiated-1_0.html#RPLogout

**Proposal**
Make the `id_token_hint` parameter optional on our logout endpoint.