🪠 Replace Flow Steps with session validator
**Problem**
Flow Steps were an attempt to ensure that no unexpected errors occur in
a controller.
However, instead of authorizing or blocking specific page transitions,
we would rather redirect the user back to the start of the intersection
(when possible).

**Proposal**
Routing errors are split into two categories:
- There is an ongoing `oidc-provider` interaction:
If we land on a step with missing session data, we can safely return to
the start of the interaction.
- There is no ongoing `oidc-provider` interaction:
In this case, we show a page explaining that the user must return to the
service they came from, as we can no longer redirect them ourselves.