✨ introduce `session.clear` function
**Problem**
We need a way to reset the session when we got a
`InvalidSessionException` to ensure we do not keep invalid data before
the redirection.
We do not want to generate a new session id on every call on authorize
when a session is not valid. This causes race condition when calling
`/authorize` twice in a row for example.

**Proposal**
Create a `clear` function that remove all data in a existing session.
This fonction is called when we do not want to reuse an existing session
in the interaction controller and when we got an
`InvalidSessionException`.

Bonus: the browsingSessionId is kept through the clear function so we
can track a user journey.

Bonus2: remove session defaultData to improve readability.