🚚 Move `oidc-provider` controller
**Problem**
We want to use `UserSessionDecorator` in `oidc-provider` controller.
This decorator is declared in the app, and we don't want to make the app
a dependency of the `OidcProvider` module.

**Proposal**
Move the `oidc-provider` controller in the app alongside our other
controllers.
This allow us to inject the same services as in the rest of the app's
controllers, and it also makes sense to keep all controllers in a single
place.