🥅 Actually forbid some email domain usage
**Problem**

There is a check that blocks the use of certain email domains for some
service providers.  However, this check is only executed on the first
connection, not when reusing an existing session.

A user could bypass the restriction by first logging in to another SP
and then accessing the SP that forbids that email domain.

**Proposal**

Run the check a second time at the end of the connection flow to ensure
it has been executed at least once.