🔧 Add advanced routing options
**Problem**

We want to block users who log in with an email domain that is not
listed among the associated domains for an IdP.

Production usage shows that some domaines _are_ acceptable for
authentication with certain IdPs, but should _not_ be used for routing.

Additionally, during the release phase, we need the ability to
temporarily unblock users if too many are being blocked by a given IdP.

**Proposal**

Add a secondary list to each IdP to record accepted domains that _should
not_ be used for routing.

Add a boolean flag to enable or disable blocking for users whose email
domain is not listed.