Revert "🚑 Avoid core to crash when proxy fails"
**Problem**

Since enabling Node proxying, we’ve encountered a Node.js bug that
causes the application to crash when an HTTP request is made to an
invalid discovery URL.

See the Node.js PR: https://github.com/nodejs/node/pull/60699

To work around this issue and preserve the previous behavior, we
restored the `global-agent` library to prevent the application from
crashing.

**Proposal**

Remove the usage of `global-agent` and use the patched version of
Node.js.

**Prerequisites**

* Restore the previous environment variables in all environments (see:
  https://github.com/proconnect-gouv/infra-legacy/pull/120).