💄 Customize messages in error page link to SP
**Problem**

On the error page, the "Retour au service" button redirects back to the
SP with a fixed message.  However, for most errors, a more specific
message is already defined.

**Proposal**

Include the exception’s `error` and `error_description` parameters in
the redirect URL so they are sent back to the service provider.